gpu: Add GdkShaderOp to the shaders
Add a gsk_gpu_image_get_shader_op() that describes the shader op to
be used in the shader. For now that returns DEFAULT always.

The shader ops are then sent into the shader via the flags.

The shader adds 2 new functions gsk_texture0(pos) and gsk_texture1(pos)
which behave like texture(GSK_TEXTURE_0, pos) and
texture(GSK_TEXTURE1, pos) respectively but do different sampling
based on the op. See the previous commit for a description of that.