gl: Don't render_texture() to MOD_INVALID dmabufs
When deciding on exporting a dmabuf vs GL texture, still prefer
the GL texture when the export to dmabuf succeeds, but uses
DRM_FORMAT_MOD_INVALID as the modifier.

This is legacy code and when in legacy mode, staying with GL textures
avoids issues.

In particular, the dmabuf usage is meant for interop with Vulkan, but
Vulkan cannot deal with invalid modifiers anyway.

Avoids glitches with pre-RDNA AMD Radeon drivers.

Related: #7559