egl: Handle EGL_SWAP_BEHAVIOR
Do the following:

1. Create the EGL config with EGL_SWAP_BEHAVIOR_PRESERVED_BIT set
2. If that succeeds: EGLSetSurfaceAttrib() EGL_SWAP_BEHAVIOR to
   EGL_BUFFER_PRESERVED
3. Check EGLGetSurfaceAttrib() of EGL_SWAP_BEHAVIOR during
   get_damage_rect() and if it is not EGL_BUFFER_PRESERVED,
   report full damage and rerender the full screen.

(1) and (2) try their hardest to set EGL_BUFFER_PRESERVED but keep
working if it's not possible.

(3) does a check because even if the previous steps didn't succeed,
the surface is still free to report EGL_BUFFER_PRESERVED.

And if it doesn't work, just rerender the full screen every time.
This is equivalent to running with GSK_DEBUG=full-redraw.

Hopefully fixes some devices not working on Android

Related: #7531
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8589>