egl: Avoid SwapBuffersWithDamage() if whole buffer was damaged
This is not primarily meant as an optimization, but as a workaround
for EGL surfaces without preserved contents support.

Drivers might not like SwapBuffersWithDamage() for surfaces without
EGL_BUFFER_PRESERVED set, but we set the whole surface as damage
region.
So if when the whole surface is damaged, we avoid
SwapBuffersWithDamage(), those surfaces will never have this function
called on them.

Related: #7531