testsuite: Make colorstate test test dmabuf textures, too
This is a neat little method to ensure we exercise our cicp shader
converting *to* cicp colorstates.

The previous fixes were all discovered using this.

We do have to be a bit more lax with the image differences though
because we cannot guarantee that shaders/drivers use the same floating
point accuracy as our C code.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8603>