testsuite: Don't save compare-render output on success
If the tests succeed, we know that the output is exactly what it should
be and saving it isn't necessary.

Only do that if the test fails, or if --verbose is given.

This improves the testrunner performance and most of all avoids dumping
lots of unneeded files into CI artifacts.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8603>