testsuite: Add compare tests for subpixel offsets
This is (hopefully) an exhaustive combination of all the node
size + offset combination that a fill node can encounter. Test that for
each of those combinations we produce the right output.

These tests are about the details of the gpu renderer caching
infrastructure, so we mark them as nocairo.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8580>