gpu: Avoid pitfalls of floating point math
We assume that the bounds are snapped, (ie origin.x * sx is
integral), but computing sx * bounds->origin.x can introduce
rounding errors, which the ceil() in mod_subpixel() will then
catastrophically blow up. To avoid that, pass the actual
offset along, and use it directly.

This fixes failures of the fill-scaled-up compare test when
clipped to a small clip.

Fixes: #7572
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8547>