wayland: Plug a wl_buffer leak
Delay getting the wl_buffer until we know that we can use it
(we may not, due to color management complications). If we can't,
free the dmabuf that we may have exported a GL texture to.

This plugs a resource leak.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8599>