icon paintable: Add properties
This makes it easier to construct a correctly set up icon
paintable from ui files. This commit also changes the is-symbolic
property to be not construct-only, since that makes it be force-set
at construct time even if we don't want it to.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8647>