gsk: Fix a corner case in paths
When a subpath is started with an 'm' without a preceding 'z',
we were were forgetting to update the subpath start, causing
the next subpath to start at the wrong place.

This broke the software-installed-symbolic.svg icon in
gnome-software.

Testcase included.

Related: #7575
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8642>