gdk: Move EGL surface handling to GLContext
This moves all the code dealing with EGL from the surface
implementations into the respective EGL glcontext implementations.

The major benefit is that the surface code no longer has to even look at
OpenGL.

The major disadvantage is that it's reshuffling a lot of code, so I'd
not be surprised if bugs are lurking here.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8644>