wgl: Cache pixel format
Instead of determining the pixel format for every GL context, store
it in the display and reuse it.

This avoids a lot of work and guarantees we select the same pixel
format every time.
It also guarantees that GLContexts can be used interchangably between
GdkSurfaces, because each surface HDC is compatible with every HGLRC.

I originally wasn't sure if pixel formats can beused interchangably
between different HDCs, but the code was doing that already for the
initial context's pixel format.
And the docs for wglMakeCurrent() at
https://learn.microsoft.com/en-us/windows/win32/api/wingdi/nf-wingdi-wglmakecurrent
say:
    It need not be the same hdc that was passed to wglCreateContext when
    hglrc was created, but it must be on the same device and have the
    same pixel format.
And to me that implies interchangably of pixel formats, so I decided
this is fine.