gsk: Clarify transition hint semantics
Make the transition hint a value between 0 and 1, with 0.5 being
the mid-point that indicates linear transition. This makes it
easier to deal with non-provided values, since we can just always
set them to 0.5. And we don't need to compute the percentage at
render time either.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8655>