memoryformat: Add per-plane DXGI formats
This is useful for shader resource views, because D3D12 requires that
multiplane formats get one view per plane.

This is more or less equivalent to the multiplane support with OpenGL.

Technically, this is also equivalent to Vulkan, because it allows
mapping each plane individually, but we are using the Ycbcr extension
there which takes care of that for us.