android: Simplify surface layouting
Instead of only queuing a surface layout change in on_layout and only
applying it in compute_size, apply the layout change directly to the
surface (this *should* be legal).
This also allows us to split on_layout into two parts: onLayoutSurface
(for handling width/height/scale of surfaceChanged) and onLayoutPosition
(for when the position of the surface on the activity has changed). This
makes a code a bit nicer to understand and work with.

This patch also makes compute_size properly behaved by always calling
GdkToplevel.compute_size, not just when the surface has changed. This
should fix !7562.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8662>