Fix memory layout initialization
The new demo for snapping ran into this problem where trying
to create sub-textures at the bottom of the original texture
triggers precondition checks in gdk_memory_texture_new, even
though the math is ok and there's enough data there. The
reason is that with strides, we don't need to require the last
row to be full - it only needs to go far enough to reach the
end of the subtexture.

Note that width is in units of 'plane block bytes'.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8664>