Wayland: Properly update tablet cursors
We must clear the remembered cursor shape when we enter the
surface, so we don't erroneously avoid the set_shape call.
For good measure, do it on leave as well.

When we applied this fix for mouse pointers in 295b7efcb3abf0a,
we forgot that tablets need the same fix, since they get
proximity in/out instead of enter/leave events.

Fixes: #7593
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8674>