[Quartz] Set the view's frame in a move for temp windows.
Temp windows (context menus mostly) are created at 0,0 then moved to the
pointer location. When the move is from a Retina to a non-Retina monitor
the view needs to be redrawn at the new scale, but macOS doesn't recognize
that and the menu ends up truncated. Setting the frame on the view forces
the redraw.

Fixes https://gitlab.gnome.org/GNOME/gtk/-/issues/7598

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8678>