gtkmaplistmodel: Make 'model' property writable
Like all other models, GtkMapListModel has a 'model' property and
a setter, and is able to change the underlying model at any time.
Unlike other models, GtkMapListModel flags the 'model' property as
construct-only, which is incompatible with its own code, since the
property is clearly writable.

Remove the construct-only flag from it, and also mark the property
with the static-strings flags as well.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8696>