inspector: Mark image buttons in ui files
When explicitly setting the child property, GtkButtons heuristics
for setting the image-button or text-button style classes don't
work, we have to do it manually. Note that we are using <style>
instead of setting the css-classes property in order to not
affect other style classes such as toggle, or check. It still
doesn't work perfectly for nested cases such as GtkMenuButton.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8710>