testsuite: Run ttfautohint over our Inter subsets
The dwrite renderer has no autohinting, so if we want to have
any hope for its results to match what freetype produces, we
need to put the autogenerated hints into the font, which is
what is done here.

The rendering is unchanged on Linux, so the pngs are not
updated.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8721>