rendernodeparser: Don't abuse Pango structs
Because only 2 bits were used in the PangoGlyphVisAttr, the code tried
to repurpose the remaining bits.

That was done by casting it to an unsigned and or-ing a bit of that.

Unfortunately, sizeof(PangoGlyphVisAttr) and sizeof(unsigned) are not
guaranteed to be the same - they are on Linux, but not on Windows -
so this setting of a bit resulted in some random bit being set
elsewhere.

Fixes the text.node parser test in Windows CI.