icontheme: Fix search path order
When looking up an icon in the various locations of a spread-out
icon theme, we want to check dirs in the search path order, so
icons in ~/.local/share/icons override those in /usr/share/icons.

This was broken in 2020 in 63bbc79f4837cf6bb, when a linked
list was replaced by an array.

Fixes: #7609
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8726>