gtkstack: Fix hiding the invisible GtkStackPages on the a11y layer
The hidden state was not always set for the GtkStackPage when it was not
shown by the stack. As a result, we were trying to fetch nonsence boundary
boxes from the page's widget.

The trees resulting from that were unnecessarily complex.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8747>