Avoid conflict with preexisting modules
We are now including the print and media modules inside libgtk4.so,
and when we find modules from an earlier installed version of gtk
in the filesystem, their type names conflict, causing ugly warnings.

Avoid this by registering our types with "Builtin" appended to
their name.

The ugliness of this is confined to the gtkmodulesprivate.h header,
where we now have a copy of G_DEFINE_TYPE_WITH_CODE.

Related: #7627
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8731>