dmabufegl: Reject multiplane import if not explicitly supported
Don't try to opportunistically import planes for multiplane dmabufs.

This code is too brittle to just go for it.

Besides, almost all multiplane formats use very standard formats for
their planes, so if a driver can't even be bothered to advertise those
correctly, it's probably a mess and we don't want to have to debug that.

Fix the driver instead.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8762>