docs: Fix non-working links within gtk4 API static generated docs
The issue opened against www.gtk.org repo describes the link to a
tutorial as not being found. Implementing the recommended URL for that link
and correcting other links to other areas of the generated website that
have since failed to reach their intended link or anchor.

I can not move or reassign tickets on gitlab, so the full path to the
issue will be mentioned here.

Closes: https://gitlab.gnome.org/Teams/Websites/www.gtk.org/-/issues/111
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8764>