entry: Add accessibility to primary and secondary icons
Currently, it's not possible to access to neither the primary
nor the secondary icons on a Gtk.Entry widget using keyboard
navigation. This means that any user relying on that, or on a
screen reader, won't be able to access these icons when they
are made activatable.

In these cases, the Gtk convention is to add those same actions
in the GtkEntry context menu, to ensure that they are accessible
when using keyboard navigation and/or a screen reader.

This MR adds two new properties, `menu-entry-icon-primary-text`
and `menu-entry-icon-secondary-text`, along with the
corresponding methods `gtk_entry_get_menu_entry_icon_text()` and
`gtk_entry_set_menu_entry_icon_text()`. If the primary icon in
the GtkEntry is activatable and a text have been set in the
first property, a new menu item will appear in the context
menu with the specified text. If the user selects that item,
the GtkEntry widget will emit the same `activate` signals
that it would if the user had clicked on the primary icon.
The same for the secondary icon.

Also, if the corresponding icon is activatable but insensitive,
the menu item will still appear, but insensitive, as expected.

This greatly simplifies implementing accessibility support in
Gtk4 apps for the primary and secondary icons, because it only
requires to set the text for the menu entry, and everything
else will be managed automatically.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8774>