gpu: Don't pass structs as varying between shader stages
The Adreno Android driver has an issue when a varying struct is passed
into a subroutine, likely due to insufficient testing when porting their
driver from GLES 2.0 (where this wasn't allowed at all) to 3.0.

As a consequence, we should avoid passing any structs between shader
stages. Specifically, this means instead of a declaring Rect and
RoundedRect as structs, we declare them as vec4[1] and vec4[3]
respectivly, which seem to work on Adreno.

Fixes: #7531
Supersedes: !8784
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8789>