print: Add support for gnome-papers as previewer
Look for the gnome-papers flatpak or a host-installed binary
before falling back to evince.

This should work both for systems that have the latest GNOME
as well as for stable distros that stick to evince.

It is still possible for apps or users to override the previewer
command using the gtk-print-preview-command setting.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8796>