GtkSpinButton: Don't expect monetary positive/negative signs
Changed gtk_spin_button_insert_text to only look for '+' and '-'
to address #7683 (disappearing negative values):

According to POSIX, the `positive_sign` and `negative_sign` fields in
struct lconv are meant for monetary quantities. They are not produced
by sprintf which formats the spinbutton (and can't be converted to a
number by strtod if non-ASCII).

This caused the spinbutton to be unable to accept its own formatting
in locales where the monetary negative_sign was something else than
ASCII '-'.

Additionally, (minor) setting FALSE/TRUE instead of 0/1 to the `sign`
boolean.

Fixes #7683

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8802>