print portal: fix portal->destroy invocations for the asynchronous case
portal->destroy must not be called in finish_response as portal is
still used afterwards. So, the right place to put it is the end of
prepare_print_response. Besides, it must be called if and only if g_main_loop_quit
is called on portal->loop (synchronous case), so this also fixes
a leak in prepare_print_called.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8815>