adjustment: Verify that we have a valid range
While implementing a scrollable widget, I made the mistake of
configuring an adjustment like so:

    gtk_adjustment_configure (adj, value, 0, content_size,
                              viewport_size * 0.1, viewport_size * 0.9,
                              viewport_size);

This almost works; however this results in subtle breakage (overshoots
always being visible in GtkScrolledWindow) when content_size is less
than viewport_size. And indeed, the actual range of the values of an
adjustment is from 'lower' up to 'upper - page_size', so configuring one
with 'lower = 0' and 'upper < page_size' is simply invalid.

We cannot catch this when setting lower/upper/page_size individually,
since they may be set in any order. Fortunately, we can catch it in
gtk_adjustment_configure, which is commonly used to set all these
properties at the same time, as well as in the constructor.

Also, fix a test that was actually violating this invariant.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8824>