gdk/wayland: Support wl_seat version 10
This commit handles wl_pointer.axis_relative_direction (although
it does nothing with it, yet), and the "repeated" key state in
wl_keyboard.key events (implementing keyboard repeat completely
on the compositor side).

This fixes the slow keys a11y option in GTK clients, as this
option cannot be expressed in terms of wl_keyboard.repeat_info,
and client-side keyboard repeat kicked in in undesired situations.

As a consequence, the dependency on wayland core was bumped to
1.24.0.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8846>