gdk: Add GdkScrollRelativeDirection enum field in scroll events
This enum will express whether the scroll event deltas/direction
matches the physical motion on the device, or whether it is inverted
relative to the physical motion.

In some devices and backends, we cannot know this, so an additional
"unknown" value is observed.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8836>