gtkrange: Only listen to horizontal scroll on surface scroll units
Even though wheels are capable of horizontal discrete scrolling, it is
still quite expected that vertical scroll (i.e. turning the wheel, instead
of sideways clicks) works there, as most mice with wheels do not have
horizontal scroll.

It is on other scrolling methods, such as touchpads and a designated
scroll button, that we can guarantee horizontal scrolling, and expect
that with horizontal ranges.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8836>