gtkrange: Accumulate scroll deltas on devices capable of begin/end hints
If we are able to get begin/end hints of scroll action, store the initial
value of the range, and operate on scroll deltas accumulated from that
point.

This helps for discrete GtkScales having marks that are far apart enough
that a single scroll event is not sufficient to move between them. Barely
noticeable for now, since scroll deltas are mapped directly to
GtkAdjustment values, but will become more noticeable in later commits.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8836>