Correct license of GtkFileThumbnail sources
The text of the license says version 3.0 of LGPL. The SPDX
identifier says LGPL-2.0-or-later. I think Georges actually intended to
specify LGPL-2.1-or-later, since this is the license that applies to
most of GTK. LGPL-2.0-or-later would be fine too, but that is notably
the GNU *Library* General Public License, not the GNU Lesser General
Public License, so the text would need to be adjusted even more to match
that.

I will ask Georges to approve this change on GitLab before
landing.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8844>