Delete Kikongo and Turkmen translations
Checking Damned Lies, I see these translations have 3% and 1% UI
coverage for GNOME 48, respectively, so there's no value in keeping them.

The Turkmen translation is annoying because it is licensed under GPL
without specifying which version. I interpret this to mean
GPL-1.0-or-later, which makes GTK's already very complex license a
little more complicated. See
https://src.fedoraproject.org/rpms/gtk4/pull-request/21 for full license
analysis.

The Kikongo translation is less annoying since it is licensed
LGPL-2.0-or-later, which is a normal license for GTK, and does not
affect GTK's overall license.  But it seems nice to clean up
translations that are not licensed under the same license as the project.

https://gitlab.gnome.org/Teams/Translation/Coordination/-/issues/361

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8848>