gdk/wayland: Flush all tablet events after wp_tablet_tool.frame
So far, we have only accumulated one type of tablet event to be emitted
and propagated during the wp_tablet_tool.frame event, and emitted any
intermediate accumulated events with the time of the last event frame.

This may have misleading effects down the line in some devices. Namely,
those tablets without proximity, as we would try to send both proximity
in and button press (or, button release and proximity out) in a rapid
sequence, one of them with a timestamp in the past.

Instead, accumulate all these events, and give them a proper event time
as given by wp_tablet_tool.frame. These are all flushed in order of
arrival with a meaningful evtime.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8861>