build: pass output file as argument to docutils rst2man/html
It has always supported passing the output file as a parameter.
Previously, we used a crude hack of having meson launch a second program
with a longer command line simply to soak up stdout of rst2man and write
it to a file. But simply telling the tool to do so directly is both
simpler and faster.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8858>