GdkWin32Display: Add entry for Intel GPUs
This adds an entry for Intel GPUs in GdkWin32Display so that we fallback
to the Cairo renderer if an Intel GPU is being used. Nothing is done yet
for this, and a getter for this entry is added.

Intel drivers aren't very capable of handling D3D12 rendering well,
especially when video playback is being involved.