Add GtkPathPaintable
This is a paintable implementation that supports a subset
of SVG similar to symbolic icons, with some extensions to
allow transitions and animations.

This commit includes documentation and a mime type for
the new icon format.

No attempt is made to integrate this with the icon theme
machinery. The expected usage is to add icon files as
resources, and then create path paintables in ui files.

There is currently no automatic hook-up of states.
Property bindings can be used for that.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8892>