dnd: Be more proactive about cleaning up
When the drop signal handler returns, the drop is done.
At that point, we can clean out the drop object and reset
our style, without waiting for a leave event (that might
not come).

Also, clean out a possibly lingering drop object when the
controller is finalized, to prevent leaks.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8894>