columnview: Fix arrow navigation in a nested widget inside a cell
If a column view cell gets a complex widget with which should be arrow keys
navigable, the previous implementation did not pass the fact to it.

This fixes, for example, navigating inside a comtext menu inside a column
view cell.

Fixes #7345.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8897>