wayland: Invert flipped transforms
So the visual results match when running with and without
GDK_DISABLE=offload. The comment above already hinted in this direction,
without being matched by the code - it looks like GdkDihedral applies
flips and rotations in the inverse manner.

This came up when testing the upcoming versions of Snapshot, GTK and the
gtk4paintablesink on mobile phones using the libcamera software ISP. In
this scenario there are usually two cameras, both 90 or 270 degree
rotated - and the front one will get an additional flip in order to
create a "mirror" effect. The recent improvements to the mentioned
components resulted in offloading now being used by default (the
software ISP generates RGB buffers - thus unlike for YUV formats
GDK_DEBUG=color-mgmt is not required for offloading).

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8931>