Portal check: activate the service if needed
According to the xdg-desktop-portal system integration docs [1],
xdg-desktop-portal and its backends are activatable DBus services. This
implies that clients cannot assume that the portals are already running,
and should instead make use of DBus activation to start the portals on
first use. Indeed, even in GNOME the portals are not started immediately
at login, and instead are started on-demand when first used shortly
afterwards.

There is, however, a gap in time where the session is running but the
portals have not yet been requested. At this moment, apps auto-started
by gnome-session are launched. As GTK queries for the presence of
various portals (with the NO_AUTOSTART flag set) it will conclude that
no portals are available on the system. This causes GTK to misbehave, as
it avoids using the portals.

This patch allows GTK to autostart the portals (which corrects the
behavior of the portal presence detection) without unnecessarily waiting
for a timeout on systems where portals are unavailable.

Follow-up for 4df849ded5aa1130fc0193aac40f50b74ca1314c

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8967>