gtkapplication-wayland: Simplify chainup
GtkApplicationImplWayland stores the parent vfuncs in its own class
struct, and uses that as a way to "chain up" to these vfuncs. This
is a rather unusual approach, and we can simply use the parent
type's class struct for that.

This unexpected approach is probably what caused the regression from
commit 5ec1fdff1d6.

Use the common ways to chain up instead of storing the parent vfuncs
in its own class.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/8972>