quartz: gdkselection check compile time dependency with MAC_OS_X_VERSION_MAX_ALLOWED
Use MAC_OS_X_VERSION_MAX_ALLOWED instead of MAC_OS_X_VERSION_MIN_REQUIRED
to check if we can currently compile in the functions from
UniformTypeIdentifiers. If the sdk supports it, we compile it in.
The runtime check takes a different path if we run on a system that
does not support UniformTypeIdentifiers functions, i.e if we run
on MacOS < 11.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/7332>