vulkan: Use the factored out functions in the dmabuf path
This is also meant for refactoring, but it omits the size check that the
code did previously and doesn't set mipmap flags.

This is shouldn't be relevant (if we have an image it's size will be
supported and it'l break later if it doesn't) or is actually a bugfix
(we never want to advertise CAN_MIPMAP for readonly images).

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9056>