gtkmenu: Await more motion events before deactivating on release
Under some circumstances, it is possible to get into situations where
a GtkMenu is popped down immediately on button release even on short
clicks. The circumstances that exhibit this behavior are:

- The GtkMenu has to pop up underneath the pointer. E.g. comboboxes,
  but not necessarily just that.
- The compositor has to send a motion event between button press and
  release. E.g. RDP and EIS do that, in order to ensure the button
  release happens in the right coordinates. But this might as well
  happen in other circumstances, input devices and compositors.

This results in GtkMenu thinking that this motion event was triggered
by user input, thus deeming it no longer necessary to keep the menu
opened after a short press+release.

While this behavior makes some sense, doing so on the first synthetic
motion event received is a bit too eager, so fix this by expecting
a (low) number of motion events before considering that an item was
definitely selected. Fixes these quickly dismissing popups in
comboboxes (and other places) when interacting through RDP.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9022>