gtkwindow: Clear active state on disappearing pointer foci
If the pointer focus is considered to be no longer on the surface
while there is an implicit grab, we were leaving the active state
stale.

On Wayland this may happen in multiple legit circumstances, e.g.
if the input is stolen by the compositor when a pointer button is
pressed, as well as other less legit situations (crossing within
surfaces from the same client). Either way, if the GtkPointerFocus
is gone, we should consider its previous active state as no longer
applying.

Closes: https://gitlab.gnome.org/GNOME/gtk/-/issues/7414
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9028>