vulkan: Refactor dmabuf feature detection code
Just like in the previous commit, return the VkFormatFeatureFlags.

We return the intersection for all supported modifiers, hoping that that
is good enough. If it weren't, we'd have to query the flag for the
created image's modifier because image creation does the modifier
selection.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9056>