x11: Restore the logic for tracking compositor presence
When a compositor is present, it claims ownership of the _NET_WM_CM_S%d
selection, where %d is the screen number. GDK has the logic to query
ownership of this selection, and set GdkDisplay:composited accordingly,
on startup. It also has the logic to request notifications for the
selection ownership changes, but no logic to handle these notifications.
Thus, GdkDisplay:composited is never updated when a compositor appears
or disappears, and a stale value is used.

GDK used to contain this logic, but it got removed in commit
437d70f56919916e884a81d3bff0170322ab2906, as a part of a larger removal
of owner change events from GDK's own API. Restore the logic, adapted to
the changes from commit 6e935d469a9f93d2b86ddec44216f1685cea92e5.

Fixes: 437d70f56919916e884a81d3bff0170322ab2906
Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>
Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9164>