testsuite: Add YUV format tests for YUYV
YUYV is a rarely used format, but not an unused one. Various webcams use
it.

It's also special because it's the only subsampled format that uses a
single channel.

It seems the Intel Vulkan driver cannot handle this format. I checked
that lavapipe can deal with it (and GL can) and I made sure there are
no validation layer complaints left, but the resulting images still look
wrong on Intel.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9056>