gdk: Remove GDK_MEMORY_U16 depth
unsigned 16bit is not something we want to deal with, we're going float
for higher depths.

The previously 16bit depth formats are now considered 32bit float
formats, because float16 only has an 11bit mantissa, so converting to
float16 would lose accuracy.

For that reason I mapped the 12bit YUV formats to float32, too.
But the 10bit YUV formats get mapped to float16, because the mantissa is
big enough there.

I'm not sure if that's the right fix and we should map things to float16
more aggressively, but I had to decide something.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9060>