testsuite: Fix alignment handling in testutils
We need to respect gdk_memory_alignment() when creating fudged buffers,
because we use internal APIs to access those buffers and those internal
APIs require the alignment.

For public API usage this already happens because we copy the memory if
necessary.

This broke the mipmap test on sparc64.

Fixes #7801

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9085>