application: Rework restore-window no-op handling
It's perfectly valid for an app to decide that it doesn't want to
restore all or some of its windows, based on the restore reason. For
instance: a multi-window app might quit with multiple windows open, and
for a normal launch may choose to restore only the first window.

So first GTK shouldn't warn about this situation, since it's valid
behavior. Second, GTK should make sure to notify the compositor that the
window won't be restored anymore, so that the compositor can clean up
its own state tracking.

Part of https://gitlab.gnome.org/GNOME/gtk/-/issues/7820

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9099>