gdk: do not run gdk_wayland_window_export_handle() callback when…
… window is already destroyed.

This is technically already handled by calling
gdk_wayland_window_unexport_handle() the right number of times as
indicated in docs strings. Yet since the case is easy to detect, and is
crash-happy, let's be a bit more robust.

Rather than just calling the exported callback, double check that the
window is not already destroyed. If it is, still do the cleanups and
output a WARNING to make sure developers of calling code are aware of
the problem in their code.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9098>