replay: Add a font filter function
This is a function that allows a simple way to filter/process just the
fonts inside a node tree.

This can be used to turn all fonts to supported fonts or convert them to
different OS's (like turn Windows fonts to equivalent Linux ones).

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9055>