application: Implement autosave
This lets applications use their save/restore support to become robust
against crashes (both of the individual app and the session). On the
startup that follows a crash, the session manager will tell the app that
is exited uncleanly (via a RECOVER restore reason) and GTK will give the
app whatever state it had saved. Periodically auto-saving their state
allows apps to lose only the last couple of seconds in such an event.

GTK takes care to only fire the autosave timer if the app is focused in
the foreground.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9099>