application: Reset restore after last window close
There are situations where the app may keep running after all of the
app's windows are gone. In this case, we want to restore state again the
next time the user tries to activate the app, because from the user's
perspective they're just relaunching the app as normal.

So, let's reset the restore reason to LAUNCH and forget that the app has
been restored already. This triggers a re-restore on the next
activation. However we don't want to restore everything: global state
doesn't need to be restored. So we're careful to not do that.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9099>