listbox: Fix GtkListBoxRow focus for vertical directions
When focusing GtkListBox from above or below, it asks the first/last row
to handle the focus. However, they just pass it through, so we end up
not focusing anything. This leads to situations like trying to go from a
header bar to a sidebar and focusing the scrolled window instead.

Fixes https://gitlab.gnome.org/GNOME/libadwaita/-/issues/892