inspector: css-editor: Replace disable button with a switch
Replace the "pause" toggle button with a switch.

A toggle button was used to disable the custom CSS, if enabled.
This "double negation" make it difficult to know if the CSS is
enabled or not.

In some case the theme can fail to differentiate and active toggle
button with it's hovered inactive state.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9107>