Fix touch dropdown selection (!8041) properly
The "fix" introduced in !8041 caused a regression when dealing with
searchable (filtered) dropdowns, due to it taking the index of the
selected item within the current filter constraints and using it as
selection index for the global model.

The proper fix is having the Gtk.ListView properly emit select-item when
it it released, regardless of its single-click-activate state. It
historically used to do this, but due to a regression introduced in
5968b10 it didn't anymore.

This reverts commit 29d2a4b0a529343ca7410c152e7a5a61d44dfbe4.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9111>