Declare an explicit dependency on zlib when broadway is enabled
When building with `broadway_enabled=true`, the gdk-3.0.pc script for
pkg-config used to declare a dependency on zlib informally, by putting
an `-lz` flag into the `Libs:` field. This causes problems when linking
applications with GTK/GDK on systems like Nix or GNU Guix, which need
explicit declarations of all dependencies and rely on pkg-config for
correctly finding them. This patch adds zlib to the `Requires` field in
gdk-3.0.pc as is appropriate for an explicit dependency declaration.

Fixes #6306

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/9119>